associations - get started, refactoring
