+++ global.d.ts
@@ -1,9 +1,9 @@
 import type { PlatformaticApp, PlatformaticDBMixin, PlatformaticDBConfig, Entity, Entities, EntityHooks } from '@platformatic/db'
-import { EntityTypes, Movie } from './types'
+import { EntityTypes, Movie,Quote } from './types'
 
 declare module 'fastify' {
   interface FastifyInstance {
-    getSchema<T extends 'Movie'>(schemaId: T): {
+    getSchema<T extends 'Movie' | 'Quote'>(schemaId: T): {
       '$id': string,
       title: string,
       description: string,
@@ -18,10 +18,12 @@ declare module 'fastify' {
 
 interface AppEntities extends Entities {
   movie: Entity<Movie>,
+    quote: Entity<Quote>,
 }
 
 interface AppEntityHooks {
   addEntityHooks(entityName: 'movie', hooks: EntityHooks<Movie>): any
+    addEntityHooks(entityName: 'quote', hooks: EntityHooks<Quote>): any
 }
 
 declare module 'fastify' {
