+++ types/index.d.ts
@@ -1,7 +1,9 @@
 import { Movie } from './Movie'
+import { Quote } from './Quote'
   
 interface EntityTypes  {
   Movie: Movie
+    Quote: Quote
 }
   
-export { EntityTypes, Movie }
+export { EntityTypes, Movie, Quote }
